PSR-4 autoloading 	- MercadoPago\Sdk namespace 	- .gitignore file 	- PhpUnit configuration file 	- Grant all tests are passing 	- Models namespace with some data helpers 	- Payment creation example
