+++ .gitignore
@@ -1,0 +1,3 @@
+vendor
+composer.lock
+.idea
