+++ examples/instant-payment-notifications/receive-ipn.php
@@ -6,7 +6,9 @@
  * @author fvaccaro
  */
 // Include Mercadopago library
-require_once "../../lib/mercadopago.php";
+require_once "../../lib/MP.php";
+
+use MercadoPago\Sdk\MP;
 
 // Create an instance with your MercadoPago credentials (CLIENT_ID and CLIENT_SECRET):
 // Argentina: https://www.mercadopago.com/mla/herramientas/aplicaciones
