+++ examples/payment-search/search-approved-payments.php
@@ -12,7 +12,8 @@
          * @author hcasatti
          */
         // Include Mercadopago library
-        require_once "../../lib/mercadopago.php";
+        require_once "../../lib/MP.php";
+        use MercadoPago\Sdk\MP;
 
         // Create an instance with your MercadoPago credentials (CLIENT_ID and CLIENT_SECRET):
         // Argentina: https://www.mercadopago.com/mla/herramientas/aplicaciones
