+++ lib/Models/Payer.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace MercadoPago\Sdk\Models;
+
+class Payer
+{
+    public $first_name = "";
+    public $last_name = "";
+    public $registration_date = "";
+    public $phone = [];
+    public $address = [];
+}
