+++ lib/Models/Shipments.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace MercadoPago\Sdk\Models;
+
+class Shipments
+{
+    public $receiver_address;
+}
